add cookie

+++ src/main/java/com/monster/zhaqsq/controller/CallListController.java
@@ -5,6 +5,8 @@ import com.github.pagehelper.PageInfo;
 import com.monster.zhaqsq.bean.CallList;
 import com.monster.zhaqsq.bean.Msg;
 import com.monster.zhaqsq.service.CallListService;
+import com.monster.zhaqsq.utils.cookieUtils;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -14,6 +16,9 @@ import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 
 @Controller
 @RequestMapping("/call")
@@ -32,20 +37,26 @@ public class CallListController {
      */
     @ResponseBody
     @RequestMapping(value = "/delete/{ids}",method = RequestMethod.DELETE)
-    public Msg deleteCall(@PathVariable("ids")String ids){
-        if (ids.contains("-")){
-            List<Integer> del_ids =new ArrayList<>();
-            String[] str_ids = ids.split("-");
-            //组装id的集合
-            for (String string : str_ids){
-                del_ids.add(Integer.parseInt(string));
-            }
-            calllistService.deleteBatch(del_ids);
-        }else {
-            Integer id =Integer.parseInt(ids);
-            calllistService.deleteCall(id);
-        }
-        return Msg.success();
+    public Msg deleteCall(@PathVariable("ids")String ids,
+			   HttpServletRequest request, HttpServletResponse response) throws Exception{
+    	if (cookieUtils.getUserType(request).equals("1")) {
+    		if (ids.contains("-")){
+    			List<Integer> del_ids =new ArrayList<>();
+    			String[] str_ids = ids.split("-");
+    			//组装id的集合
+    			for (String string : str_ids){
+    				del_ids.add(Integer.parseInt(string));
+    			}
+    			calllistService.deleteBatch(del_ids);
+    		}else {
+    			Integer id =Integer.parseInt(ids);
+    			calllistService.deleteCall(id);
+    		}
+    		return Msg.success();
+    	}
+    	else {
+			return Msg.fail();
+		}
     }
 
     /**
@@ -53,9 +64,15 @@ public class CallListController {
      */
     @ResponseBody
     @RequestMapping(value="/update/{callId}",method = RequestMethod.PUT)
-    public Msg updateCall(CallList callList){
-        calllistService.updateCallList(callList);
-        return Msg.success();
+    public Msg updateCall(CallList callList,
+			   HttpServletRequest request, HttpServletResponse response) throws Exception{
+    	if (cookieUtils.getUserType(request).equals("1")) {
+    		calllistService.updateCallList(callList);
+    		return Msg.success();
+    	}
+    	else {
+			return Msg.fail();
+		}
     }
     
     /**
@@ -63,19 +80,25 @@ public class CallListController {
      */
     @ResponseBody
     @RequestMapping(value="/updatetime",method = RequestMethod.PUT)
-    public Msg updateTime(Integer callId, String subTime, String endTime){
-    	SimpleDateFormat simpleDateFormat = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
-    	Date sTime = null,eTime = null;
-    	if(subTime!=null) {
-            long lt = new Long(subTime);
-            sTime = new Date(lt);
+    public Msg updateTime(Integer callId, String subTime, String endTime,
+			   HttpServletRequest request, HttpServletResponse response) throws Exception{
+    	if (cookieUtils.getUserType(request).equals("1")) {
+    		SimpleDateFormat simpleDateFormat = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
+    		Date sTime = null,eTime = null;
+    		if(subTime!=null) {
+    			long lt = new Long(subTime);
+    			sTime = new Date(lt);
+    		}
+    		if(endTime!=null) {
+    			long lt = new Long(endTime);
+    			eTime = new Date(lt);
+    		}
+    		calllistService.changeTime(callId, sTime, eTime);;
+    		return Msg.success();
     	}
-    	if(endTime!=null) {
-            long lt = new Long(endTime);
-            eTime = new Date(lt);
-    	}
-        calllistService.changeTime(callId, sTime, eTime);;
-        return Msg.success();
+    	else {
+			return Msg.fail();
+		}
     }
 
     /**
@@ -83,9 +106,15 @@ public class CallListController {
      */
     @RequestMapping(value = "/get/{id}",method = RequestMethod.GET)
     @ResponseBody
-    public Msg getCall(@PathVariable("id")Integer id){
-        CallList calllist = calllistService.getCall(id);
-        return Msg.success().add("call",calllist);
+    public Msg getCall(@PathVariable("id")Integer id,
+			   HttpServletRequest request, HttpServletResponse response) throws Exception{
+    	if (cookieUtils.getUserType(request).equals("1")) {
+    		CallList calllist = calllistService.getCall(id);
+    		return Msg.success().add("call",calllist);
+    	}
+    	else {
+			return Msg.fail();
+		}
     }
 
     /**
@@ -94,9 +123,15 @@ public class CallListController {
      */
     @RequestMapping(value = "/save",method = RequestMethod.POST)
     @ResponseBody
-    public Msg saveCall(CallList calllist){
-        calllistService.saveCall(calllist);
-        return Msg.success();
+    public Msg saveCall(CallList calllist,
+			   HttpServletRequest request, HttpServletResponse response) throws Exception{
+    	if (cookieUtils.getUserType(request).equals("1")) {
+    		calllistService.saveCall(calllist);
+    		return Msg.success();
+    	}
+    	else {
+			return Msg.fail();
+		}
     }
 
     /**
@@ -106,30 +141,36 @@ public class CallListController {
 
     @RequestMapping("/calls")
     @ResponseBody
-    public Msg getcallsWithJson(@RequestParam(value="pn",defaultValue = "1")Integer pn){
-        // 引入PageHelper分页插件
-        // 在查询之前只需要调用，传入页码，以及每页的大小
-        PageHelper.startPage(pn,5);
-        // startPage后面紧跟的这个查询就是一个分页查询
-        List<CallList> calls = calllistService.getAll();
-        // 使用pageInfo包装查询后的结果，只需要将pageInfo交给页面就行了。
-        // 封装了详细的分页信息,包括有我们查询出来的数据，传入连续显示的页数
-        PageInfo page = new PageInfo(calls,5);
-        return Msg.success().add("pageInfo",page);
+    public Msg getcallsWithJson(@RequestParam(value="pn",defaultValue = "1")Integer pn,
+			   HttpServletRequest request, HttpServletResponse response) throws Exception{
+    	if (cookieUtils.getUserType(request).equals("1")) {
+    		// 引入PageHelper分页插件
+    		// 在查询之前只需要调用，传入页码，以及每页的大小
+    		PageHelper.startPage(pn,5);
+    		// startPage后面紧跟的这个查询就是一个分页查询
+    		List<CallList> calls = calllistService.getAll();
+    		// 使用pageInfo包装查询后的结果，只需要将pageInfo交给页面就行了。
+    		// 封装了详细的分页信息,包括有我们查询出来的数据，传入连续显示的页数
+    		PageInfo page = new PageInfo(calls,5);
+    		return Msg.success().add("pageInfo",page);
+    	}
+    	else {
+			return Msg.fail();
+		}
     }
 
-    //@RequestMapping("/calls")
-    public String getcalls(@RequestParam(value="pn",defaultValue = "1")Integer pn, Model model){
-        // 引入PageHelper分页插件
-        // 在查询之前只需要调用，传入页码，以及每页的大小
-        PageHelper.startPage(pn,5);
-        // startPage后面紧跟的这个查询就是一个分页查询
-        List<CallList> calls = calllistService.getAll();
-        // 使用pageInfo包装查询后的结果，只需要将pageInfo交给页面就行了。
-        // 封装了详细的分页信息,包括有我们查询出来的数据，传入连续显示的页数
-        PageInfo page = new PageInfo(calls,5);
-        model.addAttribute("PageInfo",page);
-
-        return "list";
-    }
+//    //@RequestMapping("/calls")
+//    public String getcalls(@RequestParam(value="pn",defaultValue = "1")Integer pn, Model model){
+//        // 引入PageHelper分页插件
+//        // 在查询之前只需要调用，传入页码，以及每页的大小
+//        PageHelper.startPage(pn,5);
+//        // startPage后面紧跟的这个查询就是一个分页查询
+//        List<CallList> calls = calllistService.getAll();
+//        // 使用pageInfo包装查询后的结果，只需要将pageInfo交给页面就行了。
+//        // 封装了详细的分页信息,包括有我们查询出来的数据，传入连续显示的页数
+//        PageInfo page = new PageInfo(calls,5);
+//        model.addAttribute("PageInfo",page);
+//
+//        return "list";
+//    }
 }
